pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = "python django project"
authors = ["inu9431 <inu9431@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
